refactor(chat): split channel sidebar out of the inventory
inventoryComponent was overloaded: a channelMode flag switched it between the
pet-name tree and the channels sidebar (New/Join forms, per-channel menu,
bookmarks, reordering). Those channel concerns are not inventory.

inventory.js is now a mode-agnostic pet-name tree renderer that accepts an
optional `sidebar` of hooks — setupHeader, decorateItem, setupList, plus
`prepend` / `itemInitiallyHidden` flags. The channel behavior moves to
inventory/channel-sidebar.js (makeChannelSidebar), which owns the channel
reorder state and supplies those hooks; chat.js builds it in channel mode.
channelMode and the ten channel-specific options are gone from inventory's API.

Behavior-preserving (still imperative). A new baseline test,
test/component/channel-sidebar.test.js (3 tests), captures channel rendering
(New/Join header, channel-item selection, hidden non-channels, bookmarks) and
was green against the pre-refactor code; the inventory-component test (9)
covers the non-channel path.

This rescopes "whole inventory as Preact" to the pet-name tree; the channel
sidebar migrates separately later.

Verified: eslint (0 errors), prettier, tsc neutral (chat.js 30 and
inventory.js 9 pre-existing, channel-sidebar.js 0); full AVA suite 459 passing;
browser dnd 17/17 and item-actions 20/20; default-view smoke renders 16 rows
with no hidden wrappers and zero errors.
